requirements.txt
paho-mqtt==2.1.0
firebase-admin==6.5.0
python-dotenv==1.0.1
